+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.1.1] - 2020-07-13
+### Fixed
+- Documentation request couldn't handle _ in the table name correctly
+### Added
+- Missing tests which caused warnings
+
 ## [0.1.0] - 2020-07-13
 ### Changed
 - Changed the complete testing setup. (testbench)